utils: Movendo a função calcularDiferencaDias do CounterDatails.js para a pasta utils
